`Marketplace`: `Vendor` receives `Payment` upon `Order` completion (#1092)
https://github.com/zinc-collective/convene/issues/831

* Marketplace: Vendor receives Payment for Products

OK, this is the first shot across the bow for listening to stripe
checkout completions and sending the money onwards to Vendors

Co-authored-by: Dalton <daltonrpruitt@users.noreply.github.com>
Co-authored-by: Ana Ulin <anaulin@users.noreply.github.com>

* Marketplace: Register Webhook Endpoint on Stripe Account connection

I wrote a test! Aren't you proud of me, @anaulin!? Anyway, now we create
the Webhook that listens for the Stripe checkout session completed
event! Yaaaaaaaah!

* Marketplace: Persist the Stripe Webhook Secret

* Marketplace: Validate that webhook secret!!!

* Marketplace: WIP - Sketch out Request Spec for Marketplace::StripeEventsController

* Marketplace: Allow Re-Connection to Stripe for Debuggin' Purposes

We weren't sure if the webhook signing secrets were correct or not; so
we created a new webhook a few different times... Which isn't great...

So now we made it so creating an account link and a stripe webhook are
idempotent, so that they can (theoretically) be called multiple times
without triggering sadness.

* Marketplace: Request tests for splitting events

* Rooms: `last` is non-deterministic

* Marketplace: Raise an exception if we get a weird event type

---------

Co-authored-by: Dalton <daltonrpruitt@users.noreply.github.com>
Co-authored-by: Ana Ulin <anaulin@users.noreply.github.com>