Utility: Test better
We got rid of the `member-only route` shared example, because we want to trust
that pundit's raise when unscoped or un-authorized is sufficient.

We made it explicit that the `UtilityHookup` being used in `update` is a
sub-class of the UtilityHookup, which should make it less likely for us to
break support for setting sub-class attributes using the general-purpose form.

Co-authored-by: Dalton Pruitt <daltonrpruitt@users.noreply.github.com>
Co-authored-by: Ana Ulin <anaulin@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>