`Marketplace`: Prevent race-condition in `Checkout`
https://github.com/zinc-collective/convene/issues/831

So, it turns out *sometimes* Stripe will fire off that checkout session
completed event faster than we can handle it!

Now that we listen for webhook events, it makes sense that we don't
need to send people back to show their checkout; and can send them right
on to their order.

Note: We may be able to get rid of the stripe checkout session id bits
completely, since we use the transfer group as our key for looking up which
order to apply the charge to.