`Neighborhood`: Assign an `Operator` and create a `Space` on first run (#1141)
* `Neighborhood` Get rid of a "Default" Space

This gets rid of the idea of "Default" Space, in favor of a first-run
experience where an `Operator` creates a Space

* `Space`: A temporary attribute of Blueprint

OK, so maybe this should be persisted? But whatever, for now it makes sure
we can create a space via the API.

* `Neighborhood`: Allow `Authentication` at `Neibhorhood` level

Turns out, the `authenticated_session` path was implicitely a Space-only
action. With `Neighborhood` becoming the Root, we no longer can expect
sessions to only be at the Space level!

* `Neighborhood`: Provide step-by-step setup instructions

1. `Operator`s are told to configure email if they haven't already
2. `Operator`s are told how to make themselves an operator via the Rails
   console
3. Only `Operator`s may create a `Space`
4. All Links to Sign in and Out at the `Neighborhood` level work