🛠️✨ `Components`: Adds `IconButtonComponent#classes`
While attempting to use the `IconButtonComponent` for the
`Marketplace::Cart::DeliveryAddress#update` flow, we realized we didn't
have the affordances we needed to adjust the component.

So we introduced a constructor parameter for setting arbitrary classes
on the component, which makes it possible for us to adjust the
padding/text/background/etc.

🛠️🧹 `Component`: rename `IconButtonComponent`

It seems like `ButtonComponent` is a more appropriate name, given that:

- It doesn't have any `Icon` *specific* behavior and
- It's shorter

Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>