✨🌸 `Marketplace`: Tidier `Cart::DeliveryAddress` flow
- https://github.com/zinc-collective/convene/issues/831

We noticed that the `Update Address` and `Checkout` buttons were of
equivalent size/important in the visual hierarchy; which seemed a bit
wrong.

This gets rid of the bright call-to-action of the `Change Address`
button, shrinks it in size, and aligns it with the label `Deliver to:`
in the header.