🥔✨ `Marketplace`: Introduce `TaxRate#index`
- https://github.com/zinc-collective/convene/issues/1137
- https://github.com/zinc-collective/convene/commit/a53dc402402981543d13f111ed533ddf982f0c80

In a previous commit, we had tidied up the `Markeplace#edit` screen
quite a bit, but we had also removed the list of `TaxRate`s on that
screen.

This sprouts a `TaxRate#index` screen, that is linked to from
`Marketplace#edit`, so that `Member`s of a `Space` may set the `TaxRate`
for their Clients.

Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>
Co-authored-by: Dicko Sow <s12dsow@users.noreply.github.com>
Co-authored-by: Dalton Pruitt <daltonrpruitt@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>