🐞 `Marketplace`: `StripeEvent#create` was failing
- https://github.com/zinc-collective/convene/issues/1236

It looks like the `PaymentIntent` shape changes depending on what
features we request. In particular, `charges` is sometimes absent and
since we're not collecting the address from them on `Checkout` it's
not included.

It may be worth taking time to write a shot-across-the-bow feature test
that:

1. Sets up Stripe in a Space
2. Sets up the Marketplace in a Space
3. Places an Order
4. Confirms the Order was received and emails went out

But that also sounds hard.