`Marketplace`: `DeliveryWindow` gracefully handles strings
- https://zinc-collective.sentry.io/issues/4031191037/?project=6052513
- https://github.com/zinc-collective/convene/issues/1265
- https://github.com/zinc-collective/convene/issues/1185
- https://github.com/zinc-collective/convene/issues/1269

So, it turns out the reason we were getting these errors was because I
had not made the form picker smart enough to use the string value when a
delivery window was not present.

Now it is!

Also, reverting the StoreModel stuff because it did not seem to fix the
problem and I can't figure out why it's not working :'(