🛠️ `Component`: `ButtonComponent` supports `GET` Turbo Streams
- https://github.com/zinc-collective/convene/issues/1187
- https://turbo.hotwired.dev/handbook/streams#streaming-from-http-responses

It turns out that `Turbo` only does turbo-streams when
`PUT`/`POST`/`DELETE` are the request methods.

If we want to make `GET` requests use the `turbo-stream` request type,
we need to specify it in the HTML output, i.e. `<a href="..."
data-turbo-stream=true>`

I wasn't sure if this should be default behavior or not, so I leaned
towards making it default to not 🤷