`Marketplace`: Sprout `Delivery::Window` object
- https://github.com/zinc-collective/convene/issues/1185

So, the `delivery_window` field which can either be a plain string or a
particular date or even a date range is a much more sophisticated
object.

I took some time last night to explore what it could look like to use
`ActiveModel::Type`s  to create a value object which can take
responsibility for representing the data more appropriately.

That said, I think `store_model` will be better :tm:, but I didn't want
to try to dig into that just yet.