`Utilities`: Tidy up `#edit` and `#destroy` use-cases
- https://github.com/zinc-collective/convene/issues/252

I noticed my local Space had two `Stripe` utilities and I wanted to get
rid of one but I couldn't because the feature didn't exist.

So now it does.

I also ditched the `Utility` form from the `Space#edit` page and made
the list of utilities look more like the list of furniture, with edit
and trash buttons on the right.