🛠 `Neighborhood`: Adds strong_migrations gem  (#1354)
🛠 Neighborhood: Adds strong_migrations gem

- https://github.com/zinc-collective/convene/issues/1177

This change is part of an effort to add guard rails for devs writing migrations and catch *unsafe* and *illegal* operations on the database.

NOTE: There are other types of "bad" migrations that will not be caught by this gem. For example, referencing an Active Record model to execute DML. For those cases we should consider other static analysis methods like rubocop.