🧹`Components`: Pull `edit_button` up into `ApplicationComponent` (#1358)
-  https://github.com/zinc-collective/convene/issues/1187

This gives us a little tiny factory method so it's easy to drop an
`edit_button` into a page and have it look and feel very much like all
our edit buttons.

I don't know that `ApplicationComponent` is the right target for pulling
this up. Part of me wants a `ModelComponent` or something that inherits
from `ApplicationComponent` and is used in cases where the component is
for a particular `Model`, but I'm also available to wait on that until
we start seeing more stuff that wants to be grouped into such a
`Component`.