`Space`: have `Agreement`s for "Terms of Service" etc.
- https://github.com/zinc-collective/convene/issues/1154

So, I don't really know what each Space's Terms of Service, Privacy
Policy, or Content Policy will be.

I would love for us to offer a set of different terms of service
tailored for different use-cases; but for now I think it's probably best
to put the responsibility on the Space Members to make sure they have
the appropriate agreements in place.

It's kind of a weird position to take, and not one I would advocate for
if we had a space ~$30k to invest in legal, but we don't and I'm not
sure if we have any capacity to do the work ourselves or find someone
who can do it pro-bono.

So we have a very tiny `Agreement` Model which `Space`s can have many
of, and maybe I'll even get to adding a `controller` with a `#show`
action and finding a place in the UI to enumerate the
`space#agreements`.