Update owner of redis container's vol dir
No idea why 999 did work and now won't, but this 1001 does now work.
My fear is that it is due to some race condition during initialization
of the CodeSpace container or the `docker compose up` call,
but I don't have proof of either yet.
I still need to read the docs...