🥗🥔🧹✨ `Marketplace`: Test how we communicate delivery expectations
- https://github.com/zinc-collective/convene/issues/1185

As we were beginning to implement how we communicate the different
expectations based on delivery area, we noticed there were not many
tests on the current behavior; and that chunk of behavior had many many
many permutations that were meaningful.

So we took some time to pull out a `DeliveryExpectationsComponent` that
can take care of handling those permutations, and start to test them.

There is some minor wording improvements included here, specifically
because we noticed things didn't always make a ton of sense as we added
those tests; so it *does* include a bit of behavior changes.

Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Dicko Sow <s12dsow@users.noreply.github.com>