`Components`: Isolate why Lookbook is sad sometimes
OK, so from what I can tell, it is related to nested constants. LookBook
(or maybe ViewComponent?!) does not appear to follow the Zeitwerk-style
constant auto-loading rules; and rather depends on it's own internal
system for detecting changes to files and reloading the Ruby AST.

When you change a ComponentPreview that is not namespaced, it happily
reloads and everything is great. When you change a ComponentPreview that
*is* namespaced it throws a giant fit.

I have not yet created an isolated reproduction (Fresh Rails 7, Ruby
3.1, ViewComponent, etc) but I figured I would push up at least the
changes I fiddled with to figure that out last Wednesday.