🥗 `Marketplace`: Buying `Products`: System Test
- https://github.com/zinc-collective/convene/issues/1326
- https://github.com/zinc-collective/convene/issues/1657

This gives us a system test for guests buying products

As we add more things we want to test; we should be able to expand this
one test; or re-use some of the helper methods to try different
permutations.

Bear in mind, CI will need a secret token for stripe; which I should be
able to make go.