`Neighborhood`: Use a default `TimeZone` everywhere
- https://github.com/zinc-collective/convene/issues/1660

This is a small step towards a better way for handling time; since we
currently only operate out of the SF Bay Area; it seems like a
reasonable place to start.

At some point we'll want to add in time zone support for browsers,
users, spaces, etc. but this gets us to a place where things are a
little better sooner rather than later