Shorter names, and if vs unless
`unless` is kind of a mixed bag, and `standardrb` encourages us to use
`if` whenever possible.

Since we're in the `CardComponent`, it didn't feel necessary to prefix
each method with `card_`...

And I tend to like to group things with the thing they modify, rather
than their type. I.e. `footer_classes` vs `classes_footer`

But feel free to ditch this of course.