`Journal`: `Entry` can be migrated to another `Journal`
- https://github.com/zinc-collective/convene/issues/1566

Also part of the writing I'm doing on https://slipvector.games, I have a
Journal for non-canon (well, several; depending on if it's worldbuilding
or concept art or a story) and then a `Journal` for canon content.

I want to be able to take an entry and move it into the Canon journal;
and I'm fine doing it from the rails console for now; but I probably
will expose a command end-point like
`/journal/:entry_id/migrate_to?journal_id=...&keywords[]=AKeyword&keywords[]=BKeyword`
and add a button on the `Entry#edit` screen to enter the flow for
selecting the Journal / adding Keywords.