🥔✨`Journal`: Enter `Entry#new` flow via `Keyword#show`
- https://github.com/zinc-collective/convene/issues/1713
- https://github.com/zinc-collective/convene/issues/1566

I thought I would tilt at this while I waited for the build to go out;
since I noticed I go from an `Entry` to a `Keyword` when trying to
decide what to fill in next when world building.

It also auto-fills the new `Entry` body with the `Keywords` canonical
version and it's aliases!

Which... I need to make a UI for soonish.