🌸✨ `Marketplace`: `Shopper` chooses `DeliveryArea` (#1757)
* `Marketplace`: `Shopper` chooses `DeliveryArea` independently

- https://github.com/zinc-collective/convene/issues/1672

So, in reviewing a bunch of the delivery app workflows, most of them have:

- Choose Pickup or Delivery
- Enter Address
- Browse Menu
- Tap Item
- Add Item to Order

This lends me to believe we want to switch to gathering the
`DeliveryArea` *then* `DeliveryAddress` *then* showing the
`Products` that are available.

This uses TurboFrames to replace pieces of the page, rather than relying
on TurboStreams. On the one hand, it's lot easier (just respond with
html with matching `turbo_frame_tag`s`).

Blerg; anyway, I'm checking this in now; but not sure if I like it
enough.

Thoughts? In particular:

- Copy-editing
- Design tweaks?
- Name Collisions: How Undo?

* Establish more appropriate DOM Hierarchy

Card => TurboFrame => Content appears to be a little bit more reasonable
for how we nest dom things.

The Card is a layout and presentation element, so it sits as close to
the top of the dom as possible.

The Turbo Frame allows the pieces to be swapped; so it resides within
the layout.

The Content itself (be it a `DisplayArea::Form` or `DisplayArea`) is the
most inwardly nested, so that it can be replaced with frames, or laid
out differently depending on where it is being presented.

* Add tests!