🧹🛠️ `Marketplace`: Wrap the whole thing in Turbo Frames (#1793)
- https://github.com/zinc-collective/convene/issues/831

This may turn out to be a terrrrrrible idea; but it may also be fine.

Essentially, the `Marketplace` now lives within a turbo-frame; which
means it's going to be replacing itself inline. Cool? Also, horrifying.

Basically, now every `Marketplace::*Controller` will want to inherit
from `Marketplace::Controller`, so when we click a link it will
auto-magically swap out the content without requiring us to do anything
at all! Cool, huh?