`Section`: Remove `Listed` concept; since it's confusing
- https://github.com/zinc-collective/convene/issues/1155

At some point, it would make sense for us to make a Gizmo for linking to
different places (even ones outside of the Space!) and replacing the
Section Navigation bar with that.

But in the meantime, the `Listed` vs `Unlisted` dichotomy isn't really
serving us, and is getting in the way a little bit.