🐞🔨 `Section`: Deleting Sections is possible again
- https://github.com/zinc-collective/convene/issues/1841

So, this was caused by me being FANCY and putting the delete button in
the same form as the edit and update. A bad idea.

Now, deleting a Section is handled in it's own independent form, rather
than attempting to rewrite the forms hidden `_method` field to trick
Rails into doing a destroy.

This is also better from a design perspective because:

A) omg, a delete button so close to the save button is DANGEROUS
B) Deleting a Section is kinda a big deal, and while it would be better
  to bubble-wrap the heck out of it so people who *do* delete a
  section don't wind up in a sad-mad state; we can at least give them
  lots of "DANGER WILL ROBINSON DANGER" messaging. This is...
  irresponsible of us but what we can do now. If a motivated bystander
  or contributor wanted to design and implement a way for Sections to be
  Archived instead; I would be Very Happy :tm: