`Marketplace`: Very Rough  `Flyer`
- https://github.com/zinc-collective/convene/issues/1956

This gives us a QR-Code Flyer, but we should probably ditch the footer
if we can...

I sprouted a "nude" layout (Ok it needs a better name) that gives us an
empty page and breaks us out of turbo-frame land.

Anyway, to make pretty we need to:

- Attach a `Vendor`'s logo to the `Space`
- Attach a `Distributor`s logo to the `Marketplace`
- Actually work towards the fancy-pants design that Mari made!

But in the meantime, this is something that could exist and maybe even
could get better.