🥔✨ `Marketplace`: `DeliveryAreas` may be `Archived`
- https://github.com/zinc-collective/convene/issues/2014

This adds the `discard` gem so that we can archive objects. It also
updates the workflow for removing DeliveryArea to have an `Archive` step
first.

Once a `DeliveryArea` is archived, it can not yet be restored; but it
can be removed if there are no `Orders` associated with the
`DeliveryArea`