🧹 `Marketplace`: Wrap `Shopper` view, Consolidate Turbo Frames,
- While working on https://github.com/zinc-collective/convene/pull/2072,
  I started running into all kinds of weirdness because of how I had
  flailed my way to a semi-working turbo-framey experience.

Now that I have a bit better handle on things; I think we can get away
with a single TurboFrame for the Marketplace; as shown in the
`layouts/marketplace.html.erb`, and
`marketplace/_in_room_template.html.erb`.

There will still be sub-frames for when we're replacing particular
pieces of the UI (for example, when setting up `Delivery` information,
but by and large this structure appears to make things quite nice.

One side effect of this change is that the `ManagementComponent` is
always visible to `Members` of the `Space`. On the one hand; this seems
kinda not like the worst thing; since it does away with the awkward
`Manage Marketplace` button.

On the other; it means there's not really a way to "Preview" the
Marketplace other than opening a Private Browsing Window; which may not
be ideal.

Anyway, the good news is it should be a little easier to move to a browsable
`Menu` rather than our awkward CartMenuAmalgamation.