🥗 `Marketplace`: The `Square` Test is failings on CI
- https://github.com/zinc-collective/convene/issues/1500
- https://github.com/zinc-collective/convene/pull/2034

I'm not quite sure what the reason is; and didn't want to tinker too
much since @RossChapman and @AnaUlin know more about how this works.

My current theory is we need to update the [`Dependabot`
Secrets](https://github.com/zinc-collective/convene/settings/secrets/dependabot)
with the Square credentials and Locations that they set up last
Wednesday.

I think we used the operations@zinc.coop account, but the phone number
for that is not tied to the Google Voice number for
operationszinc@gmail.com; so I wasn't able to log in.

I did log in as zee@zinc.coop; and there are credentials there but I
am not brainy enough to figure out if that's the right account or not.

Anyway, to unbreak the tests for Dependabot rolling again; I am ditching
the `with_square` trait in the `marketplace` in the
`buying_products_spec.rb`. We'll probably want to write a
`marketplace/order_notifications/with_square_spec.rb` that tests that
use case more explicitely, including assertions against the Square API
that confirm we get the data in the PoS right.