🌸 `Marketplace`: `Product::Photo` tuned for iPhone
- https://github.com/zinc-collective/convene/issues/1326

The aspect ratio is now 16x9, which should be easier for people to just
"naturally" upload. The width of the small hero variant is edge-to-edge
full-def on Apple's current highest-end device, the iPhone 15 Pro Max.

There's probably a way smarter way of doing this that uses say,
background images and CSS selectors to load just-the-right-sized image
for the particular use-case; but I figure going big-and-downscaling is
probably better from a "make pretty" perspective.