🌸 `Components`: `Card` does not need a `gap-y`
- https://github.com/zinc-collective/convene/issues/2110

If we enforce a gap at the card level, it makes it hard to format things
nicely in contexts where there should not be a gap between the header
and footer.

The `justify-between` is "enough" to ensure that the footer sticks to
the bottom of the card, and the `header` and `footer` both provide
padding.

So no need to set that intention + enforce it multiple places; and this
place seemed like the one that was least tailorable to the immediate
usage context.