🥗 `Marketplace`: Test Listing `Products`
- https://github.com/zinc-collective/convene/issues/2168

It may be nice to get rid of the `Rspec/ExampleLength` cop, because omg
it doesn't make sense to apply it in the case of system specs.

But I also don't know enough about how to do that effortlessly, and I am
too lazy to put in the effort right now.

Also apparently we may be able to get rid of the `chomp`! Apparently

```rb
<<~THING.gsub("\n", " ")
  Should wind up as
  having no line breaks
THING
```

Also, it would be nice if `within(created_product)` did the same thing
as `within("##{dom_id(created_product)}")` since that looks horrifying.

Maybe `visit(object)` could do something similar with polymorphic path?

Co-Authored-By: Tim Taylor <toolbear@users.noreply.github.com>