✨ `Journal`: `Author` provides `Entry#summary` (#2161)
- When https://github.com/zinc-collective/convene-journal/issues/2, add an https://github.com/zinc-collective/convene-journal/issues/10 so we can eventually display it when https://github.com/zinc-collective/convene-journal/issues/4


* 🧹🥗 `Journal`: Add `Entry#summary` field (#8)

- https://github.com/zinc-collective/convene-journal/issues/10

What I'm driving towards here is similar to
https://github.com/zinc-collective/convene/pull/2055, where a
`Journal::Entry` has an affordance for being opinionated about how it's
represented when linked to in search engines or peer-to-peer sharing; as
well as short text for when there are a several on `Journal#show`.

* 🛠️ `Journal`: Test only the `Journal in CI

While working on
https://github.com/zinc-collective/convene-journal/pull/11, I realized I
missed a piece of the refactor because I was leaning-on-ci as a way to
confirm that everything was working as expected.

Little did I know, CI was not working in this fork. Which makes sense,
because you don't want to automatically turn on all the Workflows when
you fork a project.

This adds a Github Workflow for testing *just* the `Journal`, which
should make detecting oopsie-daisys a bit easier for folks who are only
interested in working on the Journal.

* 🥗 `Journal`: Test adding `Entry#summary` when Writing `Entries`

- https://github.com/zinc-collective/convene-journal/issues/10
- https://github.com/zinc-collective/convene-journal/issues/2

This is a quick line-of-action end-to-end test for setting an `Entry#summary`.

* ✨ `Journal`: Save `Entry#summary` when Writing `Entries`

- https://github.com/zinc-collective/convene-journal/issues/2
- https://github.com/zinc-collective/convene-journal/issues/10

And just like that, there's a `Summary` field on the
`Journal::Entry#new` page!