⚙️ `Gizmos`: Support `ActionText`
- https://github.com/zinc-collective/convene/issues/709
- https://github.com/zinc-collective/convene/pull/2200#pullrequestreview-1879198110

Markdown is cool, but it's nerd-forward.

ActionText is a much more human-friendly way of implementing rich text
in Rails applications.

I don't know enough about [Trix] to know if it will work in multi-user
contexts; or how it will work with rich-embeds, ala Notion or Google Docs

But I think it's a good-enough-default for now; and if we get to the point
where embeddables or multi-user editing of a field is important we can
figure that out.

[Trix]: https://trix-editor.org/