🌸 `Marketplace`: Improve `ManagementComponent` Iconagraphy
- https://github.com/zinc-collective/convene/issues/2220

The icons for `TaxeRates#index` and `Flyers#show` were the same, plus
`Products#index` will conflict with `Tags#index`.

So now `Products#index` is :cake: and `Flyers#show` is :qr_code:

While I was doing this, I also made the following tidyings:

- ✨ `SvgComponent` has `#building_storefront` icon!
- ✨ `SvgComponent` has `#qr_code` icon!
- 🧹 `SvgComponent` icon-method names match Heroicon's label!
- 🧹 `SvgComponent#cake` does not need a wrapping `<svg ...>` tag!
- 🥗 `ManagementComponent` has every link it's supposed to
- 🧹 `ManagementComponent::Spec` uses `have_link` instead of `have_css`
- 🧹 `ManagementComponent` eliminates usage of `ButtonComponent`

We originally added the `ButtonComponent` to inject the tailwind CSS
classes in Ruby, but with the changes @anaulin made, we now can use the
`.button` and `.button.--primary`/`.button.--secondary` classes.

Now, when I notice the `ButtonComponent` being used, I pull it out so we
can (eventually) delete it.