🧹 `Gizmo` begin ditching `Furniture`
We talked a bit at ensemble and Ana made the astute observation that
the complexity of how Gizmos work using hand-wrolled single table
inheritance is probably the biggest drag on development right now.

So we're starting to untangle that by building up a way to `Slot` in
`Gizmos` like the `Marketplace` or a `TextBlock` without the baggage of
`Furniture`'s single-table-inheritance.

Going forward, each `Gizmo` will need it's own `Model` and table for
storing data; as we will eventually ditch the `furnitures` table
completely.