WIP - Step toward Slottables
We decided that instead of creating a `Slot`, we want to create Gizmos
directly, and we want to make those `Gizmo` by sending users to the
actual `/new` page for the particular `Gizmo`.

We're thinking we will make the UI cards of Gizmos, so that they can
include things like why you might want to add that Gizmo, or the help docs
or the ... whatever. We don't know. It'll be fine.