`Marketplace`: `DeliveryArea` fees as a percentage (#2643)
- https://github.com/zinc-collective/convene/issues/2641

Todo:
- [x] A `DeliveryArea` may have a delivery fee as a percentage
- [x] `Orders` to a `DeliveryArea` with a Percentage Fee calculate the fee correctly.
- [x] `Shoppers` see a description of the Delivery Fee when building their Order.

April has adjusted her fee schedule, and catering orders are now 12% and regular orders are $10 flat. This simplifies things quite a bit; hooray! But alas, we had already built out stuff to support different fees for every vendor and across regions (i.e. SF it's $20, Oakland it's $10)

I think it's probably for the best to leave that flexibility in for now, and have updated it so we now have another attribute for how to calculate fees.

I've made the Executive Decision to treat these fees as cumulative (i.e. $10 plus 12%) if both the `DeliveryArea#price` and the `DeliveryArea#fee_as_percentage` are set; rather than trying to figure out how to break ties or validate only one is set or whatever. This may be a poor decision.